requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.45
sentencepiece>=0.2
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
Pillow>=10.0
albumentations>=1.4
opencv-python-headless>=4.10
matplotlib>=3.9
seaborn>=0.13
tqdm>=4.66
streamlit>=1.38
plotly>=5.24
folium>=0.17
pyyaml>=6.0.2
datasets>=2.20
faiss-cpu>=1.8.0
umap-learn>=0.5.6
rasterio>=1.4.0
shapely>=2.0.4
rich>=13.7.0
